+++ src/clojure/zensols/nrpuget/core.clj
@@ -6,27 +6,29 @@
   (:require [puget.printer :as puget :refer (with-options)]))
 
 (defonce ^:private options-inst (atom nil))
-(defonce ^:private options-merged-inst (atom nil))
 
 (defn set-options!
   "Set puget options as documented in the [puget
   API](https://greglook.github.io/puget/api/puget.printer.html)."
   [options]
-  (reset! options-merged-inst nil)
   (reset! options-inst options))
 
-(defn- merge-printer-opts
-  "Create optimized (merged) parameters."
+(defn option-assoc!
+  "Update only a single key/value pair rather than having to set everything
+  over again with [[set-options!]]."
+  [key value]
+  (swap! options-inst assoc key value))
+
+(defn options
+  "Get the options currently used by puget."
   []
-  (->> (merge @options-inst
-              {:width (or *print-right-margin* 72)}
-              (if *print-length* {:seq-limit *print-length*}))
-       (puget/merge-options puget/*options*)))
+  @options-inst)
 
 (defn- printer-opts
   "Get merged parameters."
   []
-  (swap! options-merged-inst #(or % (merge-printer-opts))))
+  (->> (options)
+       (puget/merge-options puget/*options*)))
 
 (defn pprint
   "Use Puget to pretty print an object using options set with [[set-options!]]."
@@ -38,4 +40,11 @@
   "Call [println](https://clojuredocs.org/clojure.core/println) with pretty
   printed data using [[pprint]]."
   [& more]
-  (pprint more))
+  (let [set-opts (options)]
+   (with-options (merge {:width (or (:width-override set-opts)
+                                    *print-right-margin*
+                                    72)}
+                        (or (:seq-limit-override set-opts)
+                            *print-length*
+                            {:seq-limit *print-length*}))
+     (pprint more))))